+++ Intern/RayCore/src/Tracer/Ray.h
@@ -13,7 +13,7 @@ class RAYX_API Ray {
         double ydir, double zdir, double en, double s0, double s1, double s2,
         double s3, double pathLength = 0, double order = 0,
         double lastElement = 0, double extraParameter = 0);
-    Ray(glm::vec3 pos, glm::vec3 dir, glm::vec4 stokes, double energy,
+    Ray(glm::dvec3 pos, glm::dvec3 dir, glm::dvec4 stokes, double energy,
         double weight, double pathLength = 0, double order = 0,
         double lastElement = 0, double extraParameter = 0);
     Ray(double* location);
